Ajudas da linguagem e das bibliotecas foram unificadas
